% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Czech_corruption}
\alias{Czech_corruption}
\title{Czech Corruption (dataset)}
\format{
igraph object
}
\source{
Data from Tomas Diviak, Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). Converted to R format in the \code{networkdata} package from David Schoch.
}
\usage{
data(Czech_corruption, package = "DF")
}
\description{
The data comes from a Czech media database called Newton Media Search and involves all major Czech newspapers for the period from 4th June 2013 to 4th June 2014.
Actors are:
Jana Nagyova, Petr Necas - former prime minister and his office chief and love affair.
Ivan Fuksa, Petr Tluchor, Marek Snajdr - deputies of ODS (conservative governing party at that time)
Ondrej Palenik, Roman Bocek, Jan Pohunek, Milan Kovanda, Lubomir Poul, Libor Grygarek - high government officials and espionage agents
Ivo Rittig, Roman Janousek, Vaclav Ryba, Tomas Hrdlicka, Jiri Toman - eminences gris, "godfathers"
DATA FORMATS: UCINET, .csv
1-mode matrix 16 x 16 person by person.
The ties are co-appearances - every time an actor was mentioned in one article together with any other actor, it is considered to be a tie.
Ties are valued on am 11 point scale, where 10 is the strongest tie (Nagyova - Necas).
All other ties were transformed by dividing the total number of co-appearances between the two actors by the value of the strongest tie, which gave the percent of the maximal tie. This percentage was then assigned an integer value from range 0 - 9 according to which tenth of percents this particular value falls into.
Example: The Fuksa - Nagyova tie reaching 50\% of the strongest tie value was assigned a value of 5. The Nagyova - Ryba tie reaching 3\% of the max value was assigned zero etc.
}
