% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Domestic_terrorist}
\alias{Domestic_terrorist}
\title{Domestic Terrorist Web Links (dataset)}
\format{
igraph object
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). Converted to R format in the \code{networkdata} package from David Schoch.
}
\usage{
data(Domestic_terrorist, package = "DF")
}
\description{
Network of hyperlinks between domestic
terrorist group websites in the United States.
DATA FORMATS: UCINET, .csv
1-mode matrix 32 x 32 website by website
Directed binary ties are based on analysis of hyperlinks between sites.
}
\references{
Zhou et al. (2005), 'US domestic extremist groups
on the web: link and content analysis, available at http://ieeexplore.ieee.org/xpl/login.jsp?tp=&arnumber=1511999&url=http\%3A\%2F\%2Fieeexplore.ieee.org\%2Fxpls\%2Fabs_all.jsp\%3Farnumber\%3D1511999
}
