% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Hijackers_911}
\alias{Hijackers_911}
\title{9/11 Hijackers (dataset)}
\format{
igraph object
}
\source{
Reconstructed at Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks).

http://orgnet.com/tnet.html

http://firstmonday.org/ojs/index.php/fm/article/view/941/863#fig4

Converted to R format in the \code{networkdata} package from David Schoch and has been expanded (overall network) by Roger Leenders
(added: flight numbers, colors according to flights, hijacker
identification in overall network).
}
\usage{
data(Hijackers_911, package = "DF")
}
\description{
Famous dataset of the terrorists involved in the 9/11 bombing
of the World Trade Centres in 2011. Data was extracted from news reports and
ties range from 'at school with' to 'on same plane'.
1-mode matrix 19 x 19 person by person of trusted prior contacts and 1-mode
matrix 61 x 61 of other associates.
Ties are undirected and binary. Relations are a mix of prior-contacts
like trained together, lived together, financial transactions, at school with,
on same flight.
}
\examples{
\dontrun{
# full network
full_net <- Hijackers_911[[1]]
plot(full_net,
vertex.label = c(V(full_net)$name[1:19], rep("", 41)), 
vertex.size = 4, 
vertex.color = V(full_net[[1]])$flight_color)

# hijacker network
hijackers <- Hijackers_911[[2]]
plot(hijackers,
     vertex.label = c(V(hijackers)$name[1:19], rep("", 41)),
     vertex.size = 4,
     vertex.color = V(hijackers)$flight_color)
}
}
\references{
Krebs, Valdis E. "Mapping networks of terrorist cells." \emph{Connections} 24.3 (2002): 43-52.
}
