% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IS_allegiance}
\alias{IS_allegiance}
\title{Islamic State Allegiances (dataset)}
\format{
igraph object
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). Converted to R format in the \code{networkdata} package from David Schoch.
}
\usage{
data(IS_allegiance, package = "DF")
}
\description{
2-mode dataset describing groups allied to
Islamic State and the countries in which they are operating
2-mode matrix 47 x 20 organizations by state, undirected binary ties.
}
