% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IS_group}
\alias{IS_group}
\title{Islamic State Group (dataset)}
\format{
list of igraph objects
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). Reconstructed from http://www.bbc.co.uk/news/world-middle-east-29052475. Converted to R format in the \code{networkdata} package from David Schoch.
}
\usage{
data(IS_group, package = "DF")
}
\description{
The data describes relationships between members of the Islamic State supplied to the BBC by IS investigation team.
1-mode matrices 56 x 56 persons by persons undirected binary ties.
The data includes three matrices with three types of relationships:
Links (relationship definition unknown)
Components & friends (former companion, close, friends, close coordination, most important companion, close relationship)
Kinship (brothers, brothers working together, young brother, brother, married to sister, son)
}
