% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Italian_gangs}
\alias{Italian_gangs}
\title{Italian Gangs (dataset)}
\format{
igraph object
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). Converted to R format in the \code{networkdata} package from David Schoch.
}
\usage{
data(Italian_gangs, package = "DF")
}
\description{
Describes Italian gang members and their nationalities. No further contextual data available.
1-Mode matrix 67x67 person by person, relations are co-membership of gangs.
Attribute data is gang member's country of origin, coded numerically.
No codebook available.
}
