% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Linux_terrorists}
\alias{Linux_terrorists}
\title{Linux Terrorists (dataset)}
\format{
list of igraph objects
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). Converted to R format in the \code{networkdata} package from David Schoch.
}
\usage{
data(Linux_terrorists, package = "DF")
}
\description{
These are two datasets about a set of terrorists and the attacks they carry out collected by the MIND Lab at University of Maryland (UMD) (http://www.mindswap.org/).
NETWORK  1-mode matrix 645 x 645 organization by organization, directed ties. Ties are co-location of terrorist attacks.
1-mode matrix 260 x 260 organization by organization, directed ties. Ties are co-located terrorist attacks by same organisation.
}
