% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Madrid_bombing_v2}
\alias{Madrid_bombing_v2}
\title{Madrid Train Bombing 2004 v2(dataset)}
\format{
singe igraph object
}
\usage{
data(Madrid_bombing_v2, package = "DF")
}
\description{
This is a single network representation of the
relations between individuals indirectly and directly associated with the Madrid
train bombing. http://en.wikipedia.org/wiki/2004_Madrid_train_bombings
1-mode network 70 x 70 person by person.
}
