% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Montagna_mafia}
\alias{Montagna_mafia}
\title{Montagna mafia (dataset)}
\format{
list of two igraph objects
}
\source{
Cavallaro, L., Ficara, A., De Meo, P., Fiumara, G., Catanese, S.,
Bagdasar, O., Song, W., & Liotta, A. (2020).
Disrupting resilient criminal networks through data analysis:
The case of Sicilian Mafia. PLoS ONE, 15(8), e0236476.
}
\usage{
data(Montagna_mafia, package = "DF")
}
\description{
Dataset collected by Cavallaro et al. (2020) derived from the
pre-trial detention order, issued by the Court of Messina's preliminary
investigation judge on March 14, 2007, which was towards the end of the
major anti-mafia operation referred to as the "Montagna Operation".

This particular investigation was a prominent operation focused on two Mafia
clans, known as the "Mistretta" family and the "Batanesi" clan.
From 2003 to 2007, these families were found to had infiltrated several
economic activities including major infrastructure works,
through a cartel of entrepreneurs close to the Sicilian Mafia.

In the datasets, each node represents an individual, whereas links
show mutual interactions among the individuals.
The Meetings dataset, captures the physical meetings among suspects.
The Phone Calls dataset refers to phone calls among individuals.

The networks are weighted since, for each pair of individuals, we recorded
a coefficient representing the number of times the pair had a meeting
(as reported by the police surveillance logs), and made a phone call
(reported in police interceptions logs).

Vertices may belong to different categories.
Some vertices represent the leaders (i.e., "bosses") or the soldiers
(i.e., "picciotti", a Sicilian word which refers to the lowest rank of the
Mafia hierarchy) of the criminal organization.
Other nodes are associated with individuals (e.g., a fruit seller or a baker)
who had one or more calls with members of a Mafia syndicate but are not
affiliated with any criminal organization.

The dataset contains a total of 144 vertices, 95 of which are active in the
meetings graph and also 95 are active in the phone graph.
The numbers are not exactly equal to the numbers reported in the paper where
the data came from. It is unclear why.
The meetings graph has 289 edges, the phone graph has 150.
Again, not exactly equal to the reported numbers in the source paper.
}
