% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Natarajan_cocaine}
\alias{Natarajan_cocaine}
\title{Cocaine Dealing Natarajan (dataset)}
\format{
igraph object
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). Reconstructed from Natarajan, M. "Understanding the structure of a drug trafficking organization: a conversational analysis." Crime Prevention Studies 11 (2000): 273-298. Converted to R format in the \code{networkdata} package from David Schoch.
}
\usage{
data(Natarajan_cocaine, package = "DF")
}
\description{
This dataset comes from an investigation into to a large cocaine trafficking organization in New York City.
1-mode matrix 28 x 28 persons by persons.
Directed valued relations are communications exchanges / flows of information. Data come from police wiretappings (transcripts of 151 telephone conversations).
}
\references{
Natarajan, M. "Understanding the structure of a drug trafficking organization: a conversational analysis." Crime Prevention Studies 11 (2000): 273-298.
}
