% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Natarajan_heroin}
\alias{Natarajan_heroin}
\title{Heroin Dealing Natarajan (dataset)}
\format{
igraph object
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). Reconstructed from Natarajan, M. (2006). Understanding the Structure of a Large Heroin Distribution Network: A Quantitative Analysis of Qualitative Data. Quantitative Journal of Criminology, 22(2), 171-192. Converted to R format in the \code{networkdata} package from David Schoch.
}
\usage{
data(Natarajan_heroin, package = "DF")
}
\description{
This dataset comes from an investigation into a large heroin trafficking organization in New York City in the 1990s.
1-mode matrix 38x38 person by person.
Directed binary relations are communications exchanges / flows of information. Data come from police wiretappings (transcripts of 151 telephone conversations).
}
\references{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). Reconstructed from Natarajan, M. (2006). Understanding the Structure of a Large Heroin Distribution Network: A Quantitative Analysis of Qualitative Data. Quantitative Journal of Criminology, 22(2), 171-192.
}
