% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Oversize_mafia}
\alias{Oversize_mafia}
\title{Oversize mafia (dataset)}
\format{
list of three igraph objects
}
\source{
Berlusconi, G., Calderoni, F., Parolini, N., Verani, M., & Piccardi, C.
(2016). Link Prediction in Criminal Networks: A Tool for Criminal Intelligence
Analysis. PLoS ONE, 11(4), e0154244. \url{doi:10.1371/journal.pone.0154244}.

The original data are available at
\url{https://search.datacite.org/works/10.6084/m9.figshare.3156067}
}
\usage{
data(Oversize_mafia, package = "DF")
}
\description{
Dataset from operation Oversize, an Italian criminal case against
a mafia group. The investigation lasted from 2000 to 2006, and targeted
suspects involved in international drug trafficking, homicides, and robberies.
The trial started in 2007 and lasted until 2009, when the judgment was passed,
and the main suspects were convicted with penalties from 5 to 22 years of
imprisonment.
Most suspects were affiliated to the 'Ndrangheta, a mafia from Calabria
with ramifications in other regions and abroad.

Contrary to most empirical studies on criminal networks,
which rely on data derived from a single source of information,
Oversize's peculiarity lies in the availability of three networks from three
judicial documents corresponding to three different stages of the criminal
proceedings: the wiretap records (WR), the arrest warrant (AW),
and the judgment (JU).
The wiretap records include all wiretap conversations transcribed by the
police and considered relevant at first glance.
The arrest warrant contains a selection of the transcripts and other relevant
information from informants and other investigative activities (e.g. physical
surveillance).
The judgment summarizes the trial and includes information from several
sources of evidence, including wiretapping and audio surveillance.

The individuals involved in illicit activities constitute the nodes of the
networks, the links indicate a relation between any two actors.
The three networks consist of 182 nodes.
}
