% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Philippines_bombing}
\alias{Philippines_bombing}
\title{Philippines Bombing (dataset)}
\format{
list of igraph objects
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks) or here http://doitapps.jjay.cuny.edu/jjatt/data.php. Converted to R format in the \code{networkdata} package from David Schoch.
}
\usage{
data(Philippines_bombing, package = "DF")
}
\description{
This is a time series that treat specific attacks as endpoints and depict the evolution of relations between individuals indirectly and directly associated with the Philippines bombing http://en.wikipedia.org/wiki/Rizal_Day_bombings
1-mode stacked matrices 16 x 16 person by person nodes, data on 11 time periods plus kinship data and tie extinguished data.
Codebook available here http://doitapps.jjay.cuny.edu/jjatt/files/Relations_Codebook_Public_Version2.pdf
}
