% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Saxena_terror}
\alias{Saxena_terror}
\title{Saxena Terror India (dataset)}
\format{
list of igraph objects
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks). Converted to R format in the \code{networkdata} package from David Schoch.
}
\usage{
data(Saxena_terror, package = "DF")
}
\description{
Data is organisation-to-organisation links of
terrorist organisations operating in the Indian State of Jammu & Kashmir.
Four 1-mode matrices persons by persons for years 2000
(5 x 5), 2001 (25 x25), 2002 (23 x 23), 2003 (18 x 18).
Undirected, binary ties are "co-occurrence"
mentions of terrorist organisations together in various sources e.g. on-line
}
\references{
Sudhir Saxena, K. Santhanam, Aparna Basu (2004),
'Application of social network analysis (SNA) to terrorist networks' in Jammu & Kashmir, \emph{Strategic Analysis} 28(1)
}
