% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Southeast_asian}
\alias{Southeast_asian}
\title{Southeast Asian Aggregated Attacks 2005 (dataset)}
\format{
list of igraph objects
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks) or here http://doitapps.jjay.cuny.edu/jjatt/data.php. Converted to R format in the \code{networkdata} package from David Schoch.
}
\usage{
data(Southeast_asian, package = "DF")
}
\description{
The Southeast Asian
Aggregate Attack Series collapses all of the individual, Indonesian cases into
a single series of relations useful for inspecting a series of behavioural and
compositional changes in one terrorist network.
1-mode stacked matrices 109x109 person by person. Kinship, friendship, acquaintanceship, time series, tie formed, tie ended.
Codebook available here http://doitapps.jjay.cuny.edu/jjatt/files/Relations_Codebook_Public_Version2.pdf
}
