% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time_since}
\alias{time_since}
\title{Time since...}
\usage{
time_since()
}
\value{
Message!
}
\description{
The Debora & Flavio (DF) package was created to track the time they have spent together. It records the dates when they started dating, their civil marriage, and their wedding ceremony. Here’s to a life filled with blessings and a future that shines even brighter for them!
}
