% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Togo}
\alias{Togo}
\title{Project Togo (dataset)}
\format{
igraph object
}
\source{
Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks), reconstructed from Morselli's book, Inside Criminal Networks http://www.springer.com/social+sciences/criminology/book/978-0-387-09525-7 book page 187, Appendix. Converted to R format in the \code{networkdata} package from David Schoch.
}
\usage{
data(Togo, package = "DF")
}
\description{
Project Togo began in February 1998 when a Toronto-based ringing operation was dismantled and one of its participants informed the police that he was previously employed by a Montreal businessman who was also active in the resale of stolen vehicles. This initial tip was corroborated soon after by a thief who had been arrested while driving a stolen vehicle. By December 1998, the Togo investigation was under way. It spanned into February 1999 and 20 cars that were destined for France, Ghana, and local buyers in southern Quebec were retrieved.
NETWORK  1-mode matrix 33 x 33 person by person. Undirected ties.
Ties are communication exchanges between criminals. Data comes from police wiretapping.
}
\references{
Morselli, C., 2009. Inside criminal networks. New York: Springer.
}
